+++ Waypoints/Managers/MaterialReplacer.cs
@@ -83,7 +83,7 @@ namespace Waypoints.Managers
                 {
                     bundleShaders = bundle.isStreamedSceneAssetBundle && bundle
                         ? bundle.GetAllAssetNames().Select(bundle.LoadAsset<Shader>).Where(shader => shader != null)
-                        : bundle.LoadAllAssets<Shader>();
+                        : bundle?.LoadAllAssets<Shader>();
                 }
                 catch (Exception)
                 {